requirements.txt
absl-py==2.2.2
accelerate==0.20.3
# bert-score==0.3.13
# certifi==2025.4.26
# charset-normalizer==3.4.1
# click==8.1.8
# cmudict==1.0.32
# contourpy==1.3.2
# cycler==0.12.1
datasets==2.13.1
evaluate==0.4.0
# filelock==3.18.0
# fonttools==4.57.0
fsspec==2023.6.0
huggingface-hub==0.15.1
# idna==3.10
# importlib_metadata==8.7.0
# importlib_resources==6.5.2
# Jinja2==3.1.6
# joblib==1.4.2
# kiwisolver==1.4.8
# MarkupSafe==3.0.2
matplotlib==3.10.1
# mpmath==1.3.0
# networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
# packaging==25.0
pandas==2.2.3
peft==0.4.0
# pillow==11.2.1
# pip==22.3.1
# pyparsing==3.2.3
# pyphen==0.17.2
# python-dateutil==2.9.0.post0
# pytz==2025.2
# PyYAML==6.0.2
# regex==2024.11.6
# requests==2.32.3
# # rouge==1.0.1
rouge-score==0.1.2
# safetensors==0.5.3
scikit-learn==1.6.1
# scipy==1.15.2
# setuptools==65.5.0
six==1.16.0
# sympy==1.14.0
# textstat==0.7.5
# threadpoolctl==3.6.0
tokenizers==0.13.3
torch==2.2.2
# tqdm==4.67.1
transformers==4.30.2
# typing_extensions==4.13.2
# tzdata==2025.2
# urllib3==2.4.0
# zipp==3.21.0
